"modified management wise scripts"

+++ School_infrastructure/Infra_Table_Report/download_blockwise_csv.py
@@ -17,6 +17,8 @@ class donwload_blockwise_csv():
         self.p = GetData()
         self.driver.implicitly_wait(20)
         self.fname=file_extention()
+        management = self.driver.find_element_by_id('nm').text
+        management = management[16:].lower().strip()
         self.driver.find_element_by_xpath(Data.hyper).click()
         self.p.page_loading(self.driver)
         p =pwd()
@@ -25,11 +27,16 @@ class donwload_blockwise_csv():
         self.p.page_loading(self.driver)
         self.driver.find_element_by_id(Data.Download_scator).click()
         time.sleep(10)
-        self.filename = p.get_download_dir() + "/"+ self.fname.sc_block()+self.p.get_current_date()+'.csv'
+        count=0
+        self.filename = p.get_download_dir() + "/"+ self.fname.sc_block()+management+'_allBlocks_'+self.p.get_current_date()+'.csv'
+        print(self.filename)
         self.p.page_loading(self.driver)
-        return os.path.isfile(self.filename)
+        if os.path.isfile(self.filename) != True:
+            print('File is not downloaded ')
+            count = count +1
+        else:
+            print('Block level file is downloaded')
+            os.remove(self.filename)
+        return  count
 
 
-
-    def remove_csv(self):
-        os.remove(self.filename)
