+++ School_infrastructure/Infra_Table_Report/download_clusterwise_csv.py
@@ -16,6 +16,8 @@ class donwload_clusterwise_csv():
     def test_clusterwise(self):
         self.p = GetData()
         self.fname = file_extention()
+        management = self.driver.find_element_by_id('nm').text
+        management = management[16:].lower().strip()
         self.driver.find_element_by_xpath(Data.hyper).click()
         self.p.page_loading(self.driver)
         p =pwd()
@@ -24,11 +26,15 @@ class donwload_clusterwise_csv():
         self.p.page_loading(self.driver)
         self.driver.find_element_by_id(Data.Download_scator).click()
         time.sleep(5)
-        self.filename = p.get_download_dir() + "/" + self.fname.sc_cluster()+self.p.get_current_date()+'.csv'
-        self.p.page_loading(self.driver)
-        return os.path.isfile(self.filename)
+        count = 0
+        self.filename = p.get_download_dir() + "/" + self.fname.sc_cluster()+management+'_allClusters_'+self.p.get_current_date()+'.csv'
+        if os.path.isfile(self.filename) != True:
+            print('File is not downloaded ')
+            count = count + 1
+        else:
+            print('Block level file is downloaded')
+            os.remove(self.filename)
+        return count
 
-    def remove_csv(self):
-        os.remove(self.filename)
