+++ School_infrastructure/Infra_Table_Report/download_schoolwise_csv.py
@@ -17,6 +17,8 @@ class school_wise_donwload():
     def test_schoolwise(self):
         self.data = GetData()
         self.fname =file_extention()
+        management = self.driver.find_element_by_id('nm').text
+        management = management[16:].lower().strip()
         self.driver.find_element_by_xpath(Data.hyper).click()
         self.data.page_loading(self.driver)
         try:
@@ -26,13 +28,17 @@ class school_wise_donwload():
             self.data.page_loading(self.driver)
             self.driver.find_element_by_id(Data.Download_scator).click()
             time.sleep(15)
-            self.filename = p.get_download_dir() + "/" + self.fname.sc_school()+self.data.get_current_date()+'.csv'
+            self.filename = p.get_download_dir() + "/" + self.fname.sc_school()+management+'_allSchools_'+self.data.get_current_date()+'.csv'
             time.sleep(10)
-            return os.path.isfile(self.filename)
+            count = 0
+            if os.path.isfile(self.filename) != True:
+                print('File is not downloaded ')
+                count = count + 1
+            else:
+                print('Block level file is downloaded')
+                os.remove(self.filename)
+            return count
 
         except exceptions.NoSuchElementException:
             print("school wise csv downloaded")
-
-    def remove_csv(self):
-        os.remove(self.filename)
